+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 3.3.0
+  * Add new stream - transfer_reversals [#216](https://github.com/singer-io/tap-stripe/pull/216)
   * Added parent relationship for the child streams  [#214](https://github.com/singer-io/tap-stripe/pull/214)
 
 ## 3.2.0